refactor: split CLI into package
